int to long long for virtual function return types.

+++ packages/ifpack/src/Ifpack_DiagonalFilter.h
@@ -154,22 +154,22 @@ public:
     return(-1.0);
   }
 
-  virtual int NumGlobalNonzeros() const
+  virtual long long NumGlobalNonzeros() const
   {
     return(A_->NumGlobalNonzeros());
   }
 
-  virtual int NumGlobalRows() const
+  virtual long long NumGlobalRows() const
   {
     return(A_->NumGlobalRows());
   }
 
-  virtual int NumGlobalCols() const
+  virtual long long NumGlobalCols() const
   {
     return(A_->NumGlobalCols());
   }
 
-  virtual int NumGlobalDiagonals() const
+  virtual long long NumGlobalDiagonals() const
   {
     return(A_->NumGlobalDiagonals());
   }
